gui/requirements.txt
Django
uwsgi
PyMySQL
mysqlclient
Pillow
mysql-connector-python
django-debug-toolbar
djangorestframework
markdown
django-filter
django-timezone-field
